split out modules from stdlib
