+++ core/src/Luna/Builtin/Prim.hs
@@ -290,12 +290,23 @@ thunkProof f (LunaThunk a) = return $ LunaThunk $ a >>= thunkProof f
 thunkProof f (LunaSusp  a) = return $ LunaSusp  $ a >>= thunkProof f
 thunkProof f a             = f a
 
+rethrowExceptions :: IO a -> LunaEff a
+rethrowExceptions a = do
+    res <- performIO $ catchAll (Right <$> a) (return . Left . show)
+    case res of
+        Left a  -> throw a
+        Right r -> return r
+
+
 instance {-# OVERLAPPABLE #-} (FromLunaData a, ToLunaValue b) => ToLunaData (a -> b) where
     toLunaData imps f = mkPrimFun $ \d -> fromLunaData d >>= toLunaValue imps . f
 
 instance {-# OVERLAPPABLE #-} (FromLunaData a, ToLunaValue b) => ToLunaData (LunaEff a -> b) where
     toLunaData imps f = LunaFunction fun where
         fun v = toLunaValue imps $ f (v >>= fromLunaData)
+
+instance ToLunaValue a => ToLunaValue (IO a) where
+    toLunaValue imps = toLunaValue imps <=< rethrowExceptions
 
 instance ToLunaData a => ToLunaValue (LunaEff a) where
     toLunaValue imps a = toLunaData imps <$> a
